+++ bin/group_list.sh
@@ -7,11 +7,19 @@ set -e -u
 # ----------------------------------------------------------------------------
 # bin/group_list.sh
 # Returns a space separated list of the group names for this project.
-# The list is extracted from the [root_file] seciton of thexrst.toml
+# The list is extracted from the [root_file] seciton of the xrst.toml
 # file for this project.
+# If there is no xrst.toml file for this project, the group list is just
+# the default group.
 #
 # sed
 source bin/grep_and_sed.sh
+#
+if [ ! -e xrst.toml ]
+then
+   echo 'default'
+   exit 0
+fi
 #
 # temp.sed
 cat << EOF > temp.sed
